LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-resources.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error


FAIL Verify FetchEvent for resources. assert_equals: integrity of Script load (url:https://localhost:9443/service-workers/service-worker/resources/dummy?test24) must be      . expected "     " but got ""
